+++ components/raftstore/src/lib.rs
@@ -3,6 +3,7 @@
 #![cfg_attr(test, feature(test))]
 #![feature(cell_update)]
 #![feature(shrink_to)]
+#![feature(div_duration)]
 
 #[macro_use]
 extern crate bitflags;
