+++ components/raftstore/src/store/fsm/store.rs
@@ -48,7 +48,7 @@ use crate::store::util::is_initial_msg;
 use crate::store::worker::{
     CleanupRunner, CleanupSSTRunner, CleanupSSTTask, CleanupTask, CompactRunner, CompactTask,
     ConsistencyCheckRunner, ConsistencyCheckTask, PdRunner, RaftlogGcRunner, RaftlogGcTask,
-    ReadDelegate, RegionRunner, RegionTask, SplitCheckTask,
+    ReadDelegate, RegionRunner, RegionTask, SplitCheckTask, SplitHubInfo,
 };
 use crate::store::DynamicConfig;
 use crate::store::PdTask;
@@ -998,6 +998,7 @@ impl RaftBatchSystem {
         importer: Arc<SSTImporter>,
         split_check_worker: Worker<SplitCheckTask>,
         dyn_cfg: Box<dyn DynamicConfig>,
+        hub_info: SplitHubInfo,
     ) -> Result<()> {
         assert!(self.workers.is_none());
         // TODO: we can get cluster meta regularly too later.
@@ -1043,7 +1044,7 @@ impl RaftBatchSystem {
             future_poller: workers.future_poller.sender().clone(),
         };
         let region_peers = builder.init()?;
-        self.start_system(workers, region_peers, builder, dyn_cfg)?;
+        self.start_system(workers, region_peers, builder, dyn_cfg, hub_info)?;
         Ok(())
     }
 
@@ -1053,6 +1054,7 @@ impl RaftBatchSystem {
         region_peers: Vec<SenderFsmPair<RocksEngine>>,
         builder: RaftPollerBuilder<T, C>,
         dyn_cfg: Box<dyn DynamicConfig>,
+        hub_info: SplitHubInfo,
     ) -> Result<()> {
         builder.snap_mgr.init()?;
 
@@ -1144,6 +1146,7 @@ impl RaftBatchSystem {
             Arc::clone(&engines.kv),
             workers.pd_worker.scheduler(),
             cfg.pd_store_heartbeat_tick_interval.as_secs(),
+            hub_info,
         );
         box_try!(workers.pd_worker.start(pd_runner));
