+++ components/raftstore/src/store/mod.rs
@@ -43,6 +43,8 @@ pub use self::snap::{
     SnapManagerBuilder, Snapshot, SnapshotDeleter, SnapshotStatistics,
 };
 pub use self::transport::{CasualRouter, ProposalRouter, StoreRouter, Transport};
-pub use self::worker::{DynamicConfig, FlowStatistics, FlowStatsReporter, PdTask, SplitInfo};
+pub use self::worker::{
+    DynamicConfig, FlowStatistics, FlowStatsReporter, PdTask, SplitHub, SplitHubInfo, SplitInfo,
+};
 pub use self::worker::{KeyEntry, LocalReader, RegionTask};
 pub use self::worker::{SplitCheckRunner, SplitCheckTask};
