+++ src/server/node.rs
@@ -23,7 +23,7 @@ use raftstore::store::fsm::store::StoreMeta;
 use raftstore::store::fsm::{ApplyRouter, RaftBatchSystem, RaftRouter};
 use raftstore::store::SplitCheckTask;
 use raftstore::store::{self, initial_region, Config as StoreConfig, SnapManager, Transport};
-use raftstore::store::{DynamicConfig, PdTask};
+use raftstore::store::{DynamicConfig, PdTask, SplitHubInfo};
 use tikv_util::config::VersionTrack;
 use tikv_util::worker::FutureWorker;
 use tikv_util::worker::Worker;
@@ -124,6 +124,7 @@ where
         importer: Arc<SSTImporter>,
         split_check_worker: Worker<SplitCheckTask>,
         dyn_cfg: Box<dyn DynamicConfig>,
+        hub_info: SplitHubInfo,
     ) -> Result<()>
     where
         T: Transport + 'static,
@@ -160,6 +161,7 @@ where
             importer,
             split_check_worker,
             dyn_cfg,
+            hub_info,
         )?;
 
         // Put store only if the cluster is bootstrapped.
@@ -337,6 +339,7 @@ where
         importer: Arc<SSTImporter>,
         split_check_worker: Worker<SplitCheckTask>,
         dyn_cfg: Box<dyn DynamicConfig>,
+        hub_info: SplitHubInfo,
     ) -> Result<()>
     where
         T: Transport + 'static,
@@ -363,6 +366,7 @@ where
             importer,
             split_check_worker,
             dyn_cfg,
+            hub_info,
         )?;
         Ok(())
     }
